new features
GetTransactionHistory, code refactored for more performance, double to decimal

+++ Infra/Dtos/Internal/TotalAmount.cs
@@ -4,6 +4,6 @@ public class TotalAmount
 {
     public DateTime Date { get; set; }
     
-    public Double Amount { get; set; }
+    public decimal Amount { get; set; }
     
 }
